Move API source files into package
This is to help encapsulate code, but also to facilitate Dockerization
